initialize static server on first use
The static server created was never shut down, preventing node from
exiting even when just including this module (and not instantiatiating a
proxy).

Commit creates the server the first time its accessed. Since server#listen()
is asynchronous, wrap the server in a promise that resolves when
the socket is bound.